figure out that meemory issue from expanding mask inside loss but have not fixed

+++ data/_main.py
@@ -9,7 +9,7 @@ def get_datasets(dataset_config: DictConfig, save=True):
     Retrieve or generate training and validation sets based on the provided configuration.
 
     The function first checks if datasets are provided in the specified directory. If they are,
-    it will load them. If not, it will generate the datasets from the requested experimental datasets.
+    it will load them. If not, it will generate the datasets from the requested source datasets.
 
     Params
     ------
@@ -19,7 +19,7 @@ def get_datasets(dataset_config: DictConfig, save=True):
         - use_these_datasets (str or None): Directory containing train and validation datasets.
             If provided, datasets will be loaded from this directory if they exist.
             If not provided or datasets don't exist in the directory, they will be generated using the other parameters.
-        - source_datasets: Path to the experimental datasets.
+        - source_datasets: Path to the source datasets.
         - num_named_neurons (int or None): Number of named neurons to include or None to include all.
         - num_worms (int or None): Number of worms to include or None to include all.
         - num_train_samples (int): Number of training samples per worm.
@@ -50,14 +50,14 @@ def get_datasets(dataset_config: DictConfig, save=True):
     smooth_data = dataset_config.smooth_data
     train_split_first = dataset_config.train_split_first
     train_split_ratio = dataset_config.train_split_ratio
-    # Some common dataset patterns were presaved for efficiency
     ### DEBUG ###
+    # Some common dataset patterns were presaved for efficiency
     all_experiment = all([(dataset in source_datasets and source_datasets[dataset]=='all') for dataset in EXPERIMENT_DATASETS])
     if all_experiment:
+        logger.info(f"Requested dataset pattern matched the presaved `combined_AllExperimental` dataset.\n"
+                    f"Setting `use_these_datasets.path ` to that directory and loading in that dataset.")
         dataset_config.use_these_datasets.path = os.path.join(ROOT_DIR, "data", "combined_AllExperimental")
-    all_synthetic = all([(dataset in source_datasets and source_datasets[dataset]=='all') for dataset in SYNTHETIC_DATASETS])
-    if all_synthetic:
-        dataset_config.use_these_datasets.path = os.path.join(ROOT_DIR, "data", "combined_AllSynthetic")
+    # TODO: Write other dataset patterns here.
     ### DEBUG ###
     # Initialize datasets
     train_dataset, val_dataset = None, None
@@ -67,7 +67,8 @@ def get_datasets(dataset_config: DictConfig, save=True):
     ), "num_named_neurons must be a positive integer or None."
     # Make log directory
     log_dir = os.getcwd()  # logs/hydra/${now:%Y_%m_%d_%H_%M_%S}
-    os.makedirs(os.path.join(log_dir, "dataset"), exist_ok=True)
+    if save:
+        os.makedirs(os.path.join(log_dir, "dataset"), exist_ok=True)
     # Control flow for loading or creating datasets
     if dataset_config.use_these_datasets.path is not None:
         # Assert that the directory exists
@@ -112,21 +113,22 @@ def get_datasets(dataset_config: DictConfig, save=True):
                 index_col=0,
                 converters={"neurons": ast.literal_eval},
             )
-        # If both train and val splits present, load them and save their info as csv
-        if train_dataset_exists and val_dataset_exists:
-            # There's no need to save the datasets again, just their information (for visualization submodule use)
-            dataset_info_train.to_csv(
-                os.path.join(log_dir, "dataset", f"train_dataset_info.csv"),
-                index=True,
-                header=True,
-            )
-            dataset_info_val.to_csv(
-                os.path.join(log_dir, "dataset", f"val_dataset_info.csv"),
-                index=True,
-                header=True,
-            )
+        # If both train and val splits present, load and return them
+        if train_dataset_exists and val_dataset_exists and save:
+            if save:
+                # There's no need to save the datasets again, just their information (for visualization submodule use)
+                dataset_info_train.to_csv(
+                    os.path.join(log_dir, "dataset", f"train_dataset_info.csv"),
+                    index=True,
+                    header=True,
+                )
+                dataset_info_val.to_csv(
+                    os.path.join(log_dir, "dataset", f"val_dataset_info.csv"),
+                    index=True,
+                    header=True,
+                )
             return train_dataset, val_dataset
-        # Otherwise create them using the experimental datasets
+        # Otherwise create them using the source datasets
         else:
             # Logging info
             if not train_dataset_exists:
@@ -170,7 +172,7 @@ def get_datasets(dataset_config: DictConfig, save=True):
                     dataset["max_timesteps"] for _, dataset in combined_dataset.items()
                 )
                 seq_len = (min_timesteps // 2) - max_num_samples - 1
-            logger.info(f"Chosen sequence length: {min_timesteps}\n")  # DEBUG
+            logger.info(f"Chosen sequence length: {seq_len}\n")  # DEBUG
             # Split the combined dataset into train and validation datasets
             created_train_dataset, created_val_dataset, dataset_info_split = (
                 split_combined_dataset(
@@ -252,13 +254,13 @@ def get_datasets(dataset_config: DictConfig, save=True):
                 torch.save(val_dataset, os.path.join(log_dir, "dataset", f"val_dataset.pt"))
             return train_dataset, val_dataset
     else:
-        # Create the datasets using the experimental datasets
-        logger.info("Creating validation and train datasets from experimental datasets.")
+        # Create the datasets using the source datasets
+        logger.info("Creating validation and train datasets from source datasets.")
         combined_dataset, dataset_info = create_combined_dataset(source_datasets, num_named_neurons)
         # Use largest seq_len that produce num. unique samples from shortest dataset
         if seq_len is None:
             max_num_samples = max(num_train_samples, num_val_samples)
-            min_timesteps = min(dataset["max_timesteps"] for _, dataset in combined_dataset.items())
+            min_timesteps = min((dataset["max_timesteps"] for _, dataset in combined_dataset.items()))
             seq_len = (min_timesteps // 2) - max_num_samples - 1
         logger.info(f"Chosen sequence length: {seq_len}\n.")  # DEBUG
         # Split the combined dataset into train and validation datasets
@@ -347,4 +349,4 @@ def get_datasets(dataset_config: DictConfig, save=True):
 if __name__ == "__main__":
     config = OmegaConf.load("configs/submodule/dataset.yaml")
     print(OmegaConf.to_yaml(config), end="\n\n")
-    dataset = get_datasets(config.dataset, save=False)
+    dataset = get_datasets(config.dataset, save=True)
